requirements.txt
# API & Backend Engine
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3

# AI Orchestration & State Machine
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20

# AI Models & Providers
langchain-groq>=0.1.0
langchain-google-genai>=1.0.0
langchain-huggingface>=0.0.1
sentence-transformers>=2.7.0

# Vector Database
qdrant-client>=1.7.0
langchain-qdrant>=0.1.0

# Data Ingestion & Live Tools
pandas>=2.1.4
requests>=2.31.0
beautifulsoup4>=4.12.3
ddgs>=0.0.2

# User Interface
streamlit>=1.31.0

# Utilities
python-dotenv>=1.0.0